requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
opencv-python
numpy
pytest
